+++ tests/aof_test.cpp
@@ -13,6 +13,7 @@
 #include <storage/options.h>
 #include <timeutil.h>
 #include <utils/fs.h>
+#include <utils/logger.h>
 #include <utils/thread_manager.h>
 
 class Moc_Dropper : public dariadb::storage::IAofDropper {
@@ -43,6 +44,7 @@ BOOST_AUTO_TEST_CASE(AofInitTest) {
   }
 
   dariadb::utils::fs::mkdir(storage_path);
+  dariadb::utils::LogManager::start();
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storage_path, "Manifest"));
   auto aof_files = dariadb::utils::fs::ls(storage_path, dariadb::storage::AOF_FILE_EXT);
@@ -112,6 +114,7 @@ BOOST_AUTO_TEST_CASE(AofInitTest) {
   }
   dariadb::storage::Manifest::stop();
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storage_path)) {
     dariadb::utils::fs::rm(storage_path);
   }
@@ -125,6 +128,7 @@ BOOST_AUTO_TEST_CASE(AOFileCommonTest) {
   }
   {
     dariadb::utils::fs::mkdir(storage_path);
+    dariadb::utils::LogManager::start();
     dariadb::storage::Manifest::start(
         dariadb::utils::fs::append_path(storage_path, "Manifest"));
     dariadb::storage::Options::start();
@@ -141,6 +145,7 @@ BOOST_AUTO_TEST_CASE(AOFileCommonTest) {
   }
 
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storage_path)) {
     dariadb::utils::fs::rm(storage_path);
   }
@@ -153,7 +158,7 @@ BOOST_AUTO_TEST_CASE(AOFManager_Instance) {
     dariadb::utils::fs::rm(storagePath);
   }
   dariadb::utils::fs::mkdir(storagePath);
-
+  dariadb::utils::LogManager::start();
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storagePath, "Manifest"));
 
@@ -173,7 +178,7 @@ BOOST_AUTO_TEST_CASE(AOFManager_Instance) {
   dariadb::storage::Manifest::stop();
   dariadb::utils::async::ThreadManager::stop();
   dariadb::storage::Options::stop();
-
+dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }
@@ -190,6 +195,7 @@ BOOST_AUTO_TEST_CASE(AofManager_CommonTest) {
     dariadb::utils::fs::rm(storagePath);
   }
   dariadb::utils::fs::mkdir(storagePath);
+  dariadb::utils::LogManager::start();
   {
     dariadb::storage::Manifest::start(
         dariadb::utils::fs::append_path(storagePath, "Manifest"));
@@ -245,6 +251,7 @@ BOOST_AUTO_TEST_CASE(AofManager_CommonTest) {
     dariadb::utils::async::ThreadManager::stop();
     dariadb::storage::Options::stop();
   }
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }