+++ tests/pages_test.cpp
@@ -15,6 +15,7 @@
 #include <utils/fs.h>
 #include <utils/thread_manager.h>
 #include <utils/utils.h>
+#include <utils/logger.h>
 
 using dariadb::storage::PageManager;
 using dariadb::storage::Manifest;
@@ -26,6 +27,7 @@ BOOST_AUTO_TEST_CASE(ManifestFileTest) {
   }
 
   std::string version="0.1.2.3.4.5";
+  dariadb::utils::LogManager::start();
   {
     Manifest::start(fname);
     std::list<std::string> pages_names{"1", "2", "3"};
@@ -69,6 +71,7 @@ BOOST_AUTO_TEST_CASE(ManifestFileTest) {
     BOOST_CHECK_EQUAL(Manifest::instance()->get_version(), version);
     Manifest::stop();
   }
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(fname)) {
     dariadb::utils::fs::rm(fname);
   }
@@ -109,6 +112,7 @@ BOOST_AUTO_TEST_CASE(PageManagerInstance) {
     dariadb::utils::fs::rm(storagePath);
   }
   dariadb::utils::fs::mkdir(storagePath);
+  dariadb::utils::LogManager::start();
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storagePath, dariadb::storage::MANIFEST_FILE_NAME));
   dariadb::storage::Options::start();
@@ -123,6 +127,7 @@ BOOST_AUTO_TEST_CASE(PageManagerInstance) {
   dariadb::storage::Manifest::stop();
   dariadb::utils::async::ThreadManager::stop();
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }
@@ -137,6 +142,7 @@ BOOST_AUTO_TEST_CASE(PageManagerReadWriteWithContinue) {
     dariadb::utils::fs::rm(storagePath);
   }
   dariadb::Meas::MeasList addeded;
+  dariadb::utils::LogManager::start();
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storagePath, dariadb::storage::MANIFEST_FILE_NAME));
 
@@ -185,6 +191,7 @@ BOOST_AUTO_TEST_CASE(PageManagerReadWriteWithContinue) {
   dariadb::storage::Manifest::stop();
   dariadb::utils::async::ThreadManager::stop();
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }
@@ -199,6 +206,7 @@ BOOST_AUTO_TEST_CASE(PageManagerMultiPageRead) {
     dariadb::utils::fs::rm(storagePath);
   }
   dariadb::Meas::MeasList addeded;
+  dariadb::utils::LogManager::start();
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storagePath, dariadb::storage::MANIFEST_FILE_NAME));
 
@@ -270,6 +278,7 @@ BOOST_AUTO_TEST_CASE(PageManagerMultiPageRead) {
   dariadb::storage::Manifest::stop();
   dariadb::utils::async::ThreadManager::stop();
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }
@@ -284,7 +293,7 @@ BOOST_AUTO_TEST_CASE(PageManagerBulkWrite) {
   }
   dariadb::storage::Manifest::start(
       dariadb::utils::fs::append_path(storagePath, dariadb::storage::MANIFEST_FILE_NAME));
-
+dariadb::utils::LogManager::start();
   dariadb::storage::Options::start();
   dariadb::storage::Options::instance()->path = storagePath;
   dariadb::storage::Options::instance()->page_chunk_size = chunks_size;
@@ -362,6 +371,7 @@ BOOST_AUTO_TEST_CASE(PageManagerBulkWrite) {
   dariadb::storage::Manifest::stop();
   dariadb::utils::async::ThreadManager::stop();
   dariadb::storage::Options::stop();
+  dariadb::utils::LogManager::stop();
   if (dariadb::utils::fs::path_exists(storagePath)) {
     dariadb::utils::fs::rm(storagePath);
   }